+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Gathering all Issues and PRs which will be done in V12 and maybe are breaking ch
 - Configuration via forRoot(...) method https://github.com/damienbod/angular-auth-oidc-client/issues/747 | PR https://github.com/damienbod/angular-auth-oidc-client/pull/1041
 - Remove the "AuthorizedState" enum in Version 12 https://github.com/damienbod/angular-auth-oidc-client/issues/755 | PR https://github.com/damienbod/angular-auth-oidc-client/pull/1062
 - Use a different key than redirect to store redirect route when using autologin https://github.com/damienbod/angular-auth-oidc-client/issues/1060 | PR #1067
-- Returnvalue of loginwithpopup and login should be the same https://github.com/damienbod/angular-auth-oidc-client/issues/1048 | PR https://github.com/damienbod/angular-auth-oidc-client/pull/1066
+- Return value of `loginWithPopup` and login should be the same https://github.com/damienbod/angular-auth-oidc-client/issues/1048 | PR https://github.com/damienbod/angular-auth-oidc-client/pull/1066
 - How to provide client id during logoff https://github.com/damienbod/angular-auth-oidc-client/issues/932 | PR https://github.com/damienbod/angular-auth-oidc-client/pull/1069
 - urlHandler callback function parameter in LogoffRevocationService.logoff does nothing https://github.com/damienbod/angular-auth-oidc-client/issues/966 | PR https://github.com/damienbod/angular-auth-oidc-client/pull/1069
 - Convert all instances of "Authorized" to "Authenticated" https://github.com/damienbod/angular-auth-oidc-client/issues/1088 | PR https://github.com/damienbod/angular-auth-oidc-client/pull/1085
@@ -20,6 +20,11 @@ Gathering all Issues and PRs which will be done in V12 and maybe are breaking ch
 - Testing & Coverage
 - Improve documentation
 - Add configuration to disable or enable id_token expired check https://github.com/damienbod/angular-auth-oidc-client/issues/1113
+
+### 2021-06-12 Version 11.6.11
+
+- Silent renew does not always start
+  - [PR](https://github.com/damienbod/angular-auth-oidc-client/pull/1128)
 
 ### 2021-05-28 Version 11.6.10